Complete rewrite to support AWS V4 style signature
The rewrite also includes a lot of changes to the coding style. The core
logic of generating signatures now resides in a separate file from the
nginx module source. The nginx module code i.e. ngx_http_aws_auth now
only containts lines that glues the signature generation code to the
nginx lifecycle.

We also also introduced unit tests based on cmocka. Documentation on
usage of the unit tests is sparse. It however serves as a great
mechanism to both test correctness of any new functionality and also as
a way to detect regressions.